+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v1.16.5
+
+bug fix:
+
+- increase app header z index
+- show headphone when start todo
+- show activity id
+
 ## v1.16.4
 
 bug fix: